Everything uses u8 now

+++ src/board/piece.rs
@@ -31,7 +31,7 @@ impl Piece {
         }
     }
 
-    pub fn generate_moves(&self, board: &Board, row: usize, col: usize) -> Option<Vec<Move>> {
+    pub fn generate_moves(&self, board: &Board, row: u8, col: u8) -> Option<Vec<Move>> {
         match self.piece_type {
             PieceType::Pawn => generate_pawn_moves(board, row, col, self.color),
             PieceType::Bishop => generate_knight_moves(board, row, col),
